Game calibration fixed and Unity player reset added
Fixed the weird issue where unity was overwriting the start calibration value

Implemented a new way to reset player 2 from unity. We send a 2nd JSON message when we want to simulate a reset button press. This 2nd message is due to the arduino being unable to process a longer JSON message